requirements.txt
# On Raspberry Pi prefer installing OpenCV via apt for stability, but pip works for many setups.
# If pip install fails, see README for RPi-specific notes.
flask==3.0.3
gevent==24.2.1
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
imutils==0.5.4
pyyaml==6.0.2
watchdog==4.0.2
numpy<2.0
